requirements.txt
wagtail
asgiref==3.2.7
certifi==2020.4.5
chardet==3.0.4
Django==3.0.4
django-appconf==1.0.4
django-autoslug==1.9.6
django-bootstrap-modal-forms==1.5.0
django-braces==1.14.0
-e git://github.com/RaD/django-chained-selects@f8f5c467d9863e597121bdcf2535d1d6e8c3996e#egg=django_chained_select
django-classy-tags==1.0.0
django-compressor==2.4
django-contrib-comments==1.9.2
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-mptt-admin==1.0.1
django-sass-processor==0.8
django-select2==7.2.3
django-star-ratings==0.9.1
django-tag-parser==3.2
django-taggit==1.2.0
django-taggit-templatetags2==1.6.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.16
idna==2.9
libsass==0.19.4
mysqlclient==1.4.6
Pillow==7.0.0
python-akismet==0.4.1
pytz==2019.3
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
swapper==1.1.2.post1
taggit-selectize==2.7.1
unicode==2.7
Unidecode==1.1.1
urllib3==1.25.8
